move lang to utils, remove path
