+++ package.json
@@ -14,7 +14,6 @@
     "i18n": "^0.13.3",
     "input": "^1.0.1",
     "moment": "^2.29.1",
-    "path": "^0.12.7",
     "random-bigint": "0.0.1",
     "raven": "^2.6.4",
     "request": "^2.88.2",
