+++ utils/lang.ts
@@ -1,9 +1,8 @@
 import lang from 'i18n'
-import path from 'path'
 
 lang.configure({
   locales: ['en'],
-  directory: path.join(__dirname, '/locales')
+  directory: '../locales'
 })
 
 export default lang
